Refactor/#63/예외처리 개선

+++ src/main/java/com/wnis/linkyway/service/tag/TagServiceImpl.java
@@ -4,15 +4,17 @@ import com.wnis.linkyway.dto.Response;
 import com.wnis.linkyway.dto.tag.TagRequest;
 import com.wnis.linkyway.dto.tag.TagResponse;
 import com.wnis.linkyway.entity.Tag;
-import com.wnis.linkyway.exception.common.ResourceConflictException;
+import com.wnis.linkyway.exception.common.NotAddDuplicateEntityException;
+import com.wnis.linkyway.exception.common.NotDeleteEmptyEntityException;
+import com.wnis.linkyway.exception.common.NotModifyEmptyEntityException;
 import com.wnis.linkyway.repository.MemberRepository;
 import com.wnis.linkyway.repository.TagRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-
 import org.springframework.transaction.annotation.Transactional;
+
 import java.util.List;
 
 @Service("tagService")
@@ -34,14 +36,16 @@ public class TagServiceImpl implements TagService {
     
     @Override
     public Response addTag(TagRequest tagRequest, Long memberId) {
-        Tag tag = tagRepository.findByTagNameAndMemberId(tagRequest.getTagName(), memberId);
-        if (tag != null) {
-            throw new ResourceConflictException("입력 값이 이미 존재합니다.");
+        if (tagRepository.existsByTagNameAndMemberId(tagRequest.getTagName(), memberId)) {
+            throw new NotAddDuplicateEntityException("중복된 태그 이름을 추가 할 수 없습니다");
         }
-        tagRepository.addTag(tagRequest.getTagName(), Boolean.parseBoolean(tagRequest.getShareable()), memberId);
-        Tag t = tagRepository.findByTagNameAndMemberId(tagRequest.getTagName(), memberId);
+        
+        Tag tag = Tag.builder().name(tagRequest.getTagName()).shareable(Boolean.parseBoolean(tagRequest.getShareable()))
+                .views(0).member(memberRepository.getById(memberId)).build();
+        
+        tagRepository.save(tag);
         TagResponse tagResponse = TagResponse.builder()
-                .tagId(t.getId())
+                .tagId(tag.getId())
                 .build();
         
         return Response.of(HttpStatus.OK, tagResponse, "태그가 성공적으로 생성되었습니다.");
@@ -50,7 +54,7 @@ public class TagServiceImpl implements TagService {
     @Override
     public Response setTag(TagRequest tagRequest, Long tagId) {
         Tag tag = tagRepository.findById(tagId).orElseThrow(() ->
-            new ResourceConflictException("태그가 존재하지 않아 수정 할 수 없습니다.")
+                new NotModifyEmptyEntityException("태그가 존재하지 않아 수정 할 수 없습니다.")
         );
         
         tag.updateName(tagRequest.getTagName()).updateShareable(Boolean.parseBoolean(tagRequest.getShareable()));
@@ -59,9 +63,9 @@ public class TagServiceImpl implements TagService {
     
     @Override
     public Response deleteTag(Long tagId) {
-        Tag tag = tagRepository.findById(tagId).orElseThrow(()->
-            new ResourceConflictException("태그가 존재하지 않아 삭제 할 수 없습니다.")
-        );
+        if (!tagRepository.existsById(tagId)) {
+            throw new NotDeleteEmptyEntityException("태그가 존재하지 않아 삭제 할 수 없습니다");
+        }
         
         tagRepository.deleteById(tagId);
         return Response.of(HttpStatus.OK, null, "성공적으로 삭제되었습니다");
